requirements.txt
astroid==2.2.5
Django==2.0.13
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.5
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1